rip out the reset utils and move them to their own file
As you can tell by how much I moved.. yeah. Resetting shouldn't have stayed in the main program file. However.. It is also kind of broken, anyway.

Warning: untested changes.
