+++ hints/control/program.py
@@ -9,35 +9,36 @@ class Program:
     # The root folder
     root_dir = Path(getcwd())
 
+    # The root window
+    root = None
+
     # The global variable of the notepad to add to
     notebook = None
 
     # The global variable of the data tabs.
     data_tabs = {}
 
+    # The tabs that are to be created.
+    # Easily expandable later.
+    data_tab_names = [
+        'Notes',
+        'Bugs'
+    ]
+
+    # The holder of all things reset
+    resetter = None
+
     # Functions that are required elsewhere. --------------------------
+    def add_tab(self, tab_name: str) -> None:
+        '''Create a tab in the notebook.'''
+        pass
+
     def change_title(self) -> None:
         '''Change the title of the window.'''
         pass
 
-    def close_all_tabs(self) -> None:
-        '''Close all of the tabs.'''
-        pass
-
-    def close_tab(self, tab_name: str) -> None:
-        '''Close a tab in the notebook.'''
-        pass
-
-    def create_notepad_tab(self) -> None:
-        '''Recreate the primary tab.'''
-        pass
-
-    def reset_tab(self, tab_name: str, default: bool = True) -> None:
-        '''Reset the contents of the tab.'''
-        pass
-
-    def reset_tracker(self) -> None:
-        '''Completely reset the tracker.'''
+    def create_notepad(self, tab_name: str) -> CTkTextbox:
+        '''Creates a notepad under the target tab.'''
         pass
 
     def set_to_notes_tab(self) -> None:
